% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{alignVec}
\alias{alignVec}
\title{Helper function to align the content of printed tables}
\usage{
alignVec(x, digits = 3)
}
\arguments{
\item{x}{(numeric or character) vector of values to be aligned}

\item{digits}{(integer) number of digits to be displayed for numeric values}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	(character) vector of aligned values
}\if{html}{\out{</div>}}
}
\description{
This function aligns vectors by padding with spaces to the right
}
\author{
Christopher Rank \email{christopher.rank@roche.com}
}
