% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ashDensity}
\alias{ashDensity}
\title{Estimate density of distribution employing the R package "ash" using R-wrapper function.}
\usage{
ashDensity(x, ab, nbin, m, kopt = c(2, 1), normToAB = FALSE)
}
\arguments{
\item{x}{(numeric) vector of data points}

\item{ab}{(numeric) vector of lower and higher truncation limit of density estimation}

\item{nbin}{(integer) specifying the number of bins used for density estimation}

\item{m}{(integer) specifying the width of the smoothing kernel(s) used for density estimation}

\item{kopt}{(integer) vector specifying the smoothing kernel}

\item{normToAB}{(logical) specifying if the density is normed to the interval ab or to all data points in x}
}
\value{
(list) with density estimation (x values, y values, m and ab).
}
\description{
Estimate density of distribution employing the R package "ash" using R-wrapper function.
}
\author{
Christopher Rank \email{christopher.rank@roche.com}, Tatjana Ammer \email{tatjana.ammer@roche.com}
}
