% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdfTruncatedBoxCox}
\alias{cdfTruncatedBoxCox}
\title{Function to calculate the RI from a set of parameters}
\usage{
cdfTruncatedBoxCox(mu, sigma, lambda, shift, RIperc)
}
\arguments{
\item{mu}{(numeric) mean of the distribution}

\item{sigma}{(numeric) standard deviation of the distribution}

\item{lambda}{(numeric) Box-Cox transformation parameter}

\item{shift}{(numeric) shift of the distribution}

\item{RIperc}{(numeric) vector of percentiles for which the reference interval should be calculated}
}
\value{
(numeric) vector of reference interval values for the given percentiles
}
\description{
Function to calculate the RI from a set of parameters
}
\references{
Freeman, R. Modarres / Statistics & Probability Letters 76 (2006) P 767
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}, Christopher Rank \email{christopher.rank@roche.com}, Tatjana Ammer \email{tatjana.ammer@roche.com}
}
