% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{checkInvalidArgs}
\alias{checkInvalidArgs}
\title{Check Invalid Arguments}
\usage{
checkInvalidArgs(param, validArgs, marginType, verbose)
}
\arguments{
\item{param}{(list)  providing non-default parameters for the calculation of the margins}

\item{validArgs}{(character) vector of valid arguments for the calculation of the margins}

\item{marginType}{(character) specifying calculation of the margins: using "VeRUS", or Equivalence Limits ("EL")}

\item{verbose}{(logical) specifying if additional warning messages are printed}
}
\value{
warning message if invalid arguments are found
}
\description{
This function checks if the arguments in param are valid
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}
}
