% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{createVerificationTab}
\alias{createVerificationTab}
\title{Create Verification Table}
\usage{
createVerificationTab(RIperc, marginsRIdata, marginsRIcand)
}
\arguments{
\item{RIperc}{(numeric) value specifying the percentiles, which define the reference interval (default c(0.025, 0.975))}

\item{marginsRIdata}{(data.frame) Margins for the data-derived reference intervals}

\item{marginsRIcand}{(data.frame) Margins for the candidate reference intervals}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		(list) containing whether the point estimate and margin overlap tests passed, and the verification table
}\if{html}{\out{</div>}}
}
\description{
Method to summarize the UMS of RIdata and RIcand and checks of the UMs overlap
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}
}
