% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoInvHist.R
\name{defineSearchRegions}
\alias{defineSearchRegions}
\title{Helper function to define search regions for mu and sigma and to get the region around the main peak 'ab'}
\usage{
defineSearchRegions(x, lambdaVec, roundingBase, abEst = NULL)
}
\arguments{
\item{x}{(numeric) values specifying data points comprising pathological
and non-pathological values}

\item{lambdaVec}{(numeric) transformation parameter for inverse Box-Cox transformation}

\item{roundingBase}{(numeric) describing the rounding base of the dataset}

\item{abEst}{(numeric) vector with already estimated abSearchReg and abHist for second definition of search regions}
}
\value{
(list) with (abEst, search region for mu and sigma)
}
\description{
The function estimates the start search regions for mu and sigma for each lambda. Further it determines an appropriate
region around the main peak 'ab' that is used for all lambdas.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
