% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoInvHist.R
\name{estimateAB}
\alias{estimateAB}
\title{Helper function to find region around the main peak of a distribution}
\usage{
estimateAB(x)
}
\arguments{
\item{x}{(numeric) vector of data points}
}
\value{
(list) with two numeric vectors with lower and upper bound of region around the main peak used for 1) defining the search regions and
2) estimating the histogram with overlapping bins
}
\description{
Helper function to find region around the main peak of a distribution
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
