% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoInvHist.R
\name{findPeaksAndValleys}
\alias{findPeaksAndValleys}
\title{Find the index of the peaks and valleys of the density estimation.}
\usage{
findPeaksAndValleys(Dens)
}
\arguments{
\item{Dens}{(list) with density estimation (x values, y values)}
}
\value{
(list) specifying the index of the peaks and valleys of the density estimation.
}
\description{
Find the index of the peaks and valleys of the density estimation.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
