% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findRoundingBase}
\alias{findRoundingBase}
\title{Estimate rounding base of the input data.}
\usage{
findRoundingBase(x)
}
\arguments{
\item{x}{(numeric) vector of data points}
}
\value{
(numeric) with estimated rounding base (e.g. 0.001 when rounded to 3 digits)
}
\description{
Estimate rounding base of the input data.
}
\author{
Christopher Rank \email{christopher.rank@roche.com}, Tatjana Ammer \email{tatjana.ammer@roche.com}
}
