% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{getRIfromRWDRI}
\alias{getRIfromRWDRI}
\title{Get Reference Interval from RWDRI}
\usage{
getRIfromRWDRI(RWDRI, RIperc, pointEst)
}
\arguments{
\item{RWDRI}{(RWDRI) The RWDRI object containing the reference interval information.}

\item{RIperc}{(numeric) The percentiles to interpolate.}

\item{pointEst}{(character) The point estimate to use ("fullDataEst" or "medianBS").}
}
\value{
(list) A list containing the interpolated reference interval and percentiles.
}
\description{
Helper function to get the reference interval from a RWDRI object. For internal use only.
}
