% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoInvHist.R
\name{optimizeGrid}
\alias{optimizeGrid}
\title{Helper function for grid search for mu and sigma.}
\usage{
optimizeGrid(currentBestParam, paramUnique, iter, sigmLimit = TRUE)
}
\arguments{
\item{currentBestParam}{(numeric) value specifying the current best value for this parameter}

\item{paramUnique}{(numeric) vector of possible values for this parameter}

\item{iter}{(integer) indicating the number of iteration, as in the first iteration the search region
is larger than in the following iterations}

\item{sigmLimit}{(logical) specifiying if parameter is sigma and thus minimum is 0}
}
\value{
(vector) specifying the new search region fo the parameter to be optimized
}
\description{
Helper function for grid search for mu and sigma.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
