% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{plotRIVerification}
\alias{plotRIVerification}
\title{Plot method for RI verification}
\usage{
plotRIVerification(
  mar1,
  mar2,
  marginOverlap,
  Scale = c("original", "transformed", "splitXAxis"),
  lambda = 0,
  xlab = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{mar1}{(data.frame) with output of function getRIMargins()}

\item{mar2}{(data.frame) with output of function getRIMargins()}

\item{marginOverlap}{(character) vector length of mar1; specifying if and how the margins overlap; options are: "PEOverlap", "MarsOverlap", "noOverlap"}

\item{Scale}{(character) specifying if percentiles are shown on the original scale ("original") or the transformed scale ("transformed") or a split view of the x-axis ("splitXAxis")}

\item{lambda}{(numeric) specifying the power parameter (skewness) of the assumed distribution}

\item{xlab}{(character) specifying the x-axis label}

\item{title}{(character) specifying plot title}

\item{...}{(additional arguments) candLabel (character) specifying the label for the candidate RI; dataLabel (character) specifying the label for the data-derived RI}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{			(NULL) Instead, a plot is generated.
}\if{html}{\out{</div>}}
}
\description{
Plot method for RI verification
}
\author{
Christopher Rank \email{christopher.rank@roche.com}
}
