% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pnormApprox}
\alias{pnormApprox}
\title{Approximate calculation of CDF of normal distribution.}
\usage{
pnormApprox(q, pNormVal, mean = 0, oneOverSd = 1, oneOverH = 10)
}
\arguments{
\item{q}{(numeric) vector of quantiles of data points}

\item{pNormVal}{(numeric) vector of lookup table for pNorm}

\item{mean}{(numeric) vector of mean values}

\item{oneOverSd}{(numeric) reciprocal vector of sd values}

\item{oneOverH}{(numeric) defining the precision of the approximation}
}
\value{
(numeric) vector of approximate CDFs of normal distribution.
}
\description{
Approximate calculation of CDF of normal distribution.
}
\author{
Christopher Rank \email{christopher.rank@roche.com}
}
