% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{printDataFractionWithinOutsideRI}
\alias{printDataFractionWithinOutsideRI}
\title{Print Data Fractions Within and Outside Reference Intervals}
\usage{
printDataFractionWithinOutsideRI(tab, RIperc, data)
}
\arguments{
\item{tab}{(dataframe) must contain columns \code{Percentile}, \code{RICandPointEst}, and \code{RIDataPointEst}.}

\item{RIperc}{(numeric) value specifying the percentiles, which define the reference interval (default c(0.025, 0.975))}

\item{data}{(numeric) vector of unknown length, possibly containing NAs}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	(NULL) Prints the verification table to the console
}\if{html}{\out{</div>}}
}
\description{
This function prints a table showing the fractions of data points that are below, within, and above specified reference intervals.
The function handles both lower and upper limits, or just one of them, based on the input.
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}
}
