% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{printSimilarityTable}
\alias{printSimilarityTable}
\title{Helper function to print the results of getRISimilarity}
\usage{
printSimilarityTable(SimilarityTab, RIperc)
}
\arguments{
\item{SimilarityTab}{(data.frame) containing the similarity table. Must include:
\itemize{
\item Percentile: The percentile values.
\item RIdata: Reference interval data points.
\item RIcand: Reference interval candidate points.
\item max_sample_size: Maximum sample sizes.
\item s_value: S-values indicating the level of similarity.
}}

\item{RIperc}{(numeric) value specifying the percentiles, which define the reference interval}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		(NULL) Prints the similarity table to the console.
}\if{html}{\out{</div>}}
}
\description{
This function prints a formatted similarity table showing reference intervals and maximum sample sizes.
}
