% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{printTableHeader}
\alias{printTableHeader}
\title{Helper function to print Table Header}
\usage{
printTableHeader(headers, col_widths)
}
\arguments{
\item{headers}{(character) vector of column headers}

\item{col_widths}{(numeric)    vector of column widths}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		(NULL) Prints the header to the console.
}\if{html}{\out{</div>}}
}
\description{
Helper function to print Table Header
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}
}
