% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsRIVerification.R
\name{printTableRows}
\alias{printTableRows}
\title{Helper function to print Table Rows}
\usage{
printTableRows(rows, col_widths)
}
\arguments{
\item{rows}{(list) of lists containing row data.}

\item{col_widths}{(numeric) vector of column widths.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{		(NULL) Prints the rows to the console.
}\if{html}{\out{</div>}}
}
\description{
Helper function to print Table Rows
}
\author{
Matthias Beck \email{matthias.beck.mb1@roche.com}
}
