% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testcase1}
\alias{testcase1}
\title{Simulated Testcase 1.}
\format{
Numeric vector with data points.
}
\usage{
data(testcase1)
}
\description{
This dataset consists of N = 10,000 simulated measurements with 80\% non-pathological and 20\% pathological samples. 
Ground Truth for reference intervals (2.5\% perc., 97.5\% perc): [10.2, 29.8]
}
\keyword{datasets}
