% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testcase2}
\alias{testcase2}
\title{Simulated Testcase 2.}
\format{
Numeric vector with data points.
}
\usage{
data(testcase2)
}
\description{
This dataset consists of N = 50,000 simulated measurements with 60\% non-pathological and 40\% pathological samples. 
Ground Truth for reference intervals (2.5\% perc., 97.5\% perc): [59.8, 160]
}
\keyword{datasets}
