% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testcase3}
\alias{testcase3}
\title{Simulated Testcase 3.}
\format{
Numeric vector with data points.
}
\usage{
data(testcase3)
}
\description{
This dataset consists of N = 75,000 simulated measurements with 96\% non-pathological and 4\% pathological samples. 
Ground Truth for reference intervals (2.5\% perc., 97.5\% perc): [9.04, 13]
}
\keyword{datasets}
