% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testcase4}
\alias{testcase4}
\title{Simulated Testcase 4.}
\format{
Numeric vector with data points.
}
\usage{
data(testcase4)
}
\description{
This dataset consists of N = 250,000 simulated measurements with 80\% non-pathological and 20\% pathological samples. 
Ground Truth for reference intervals (2.5\% perc., 97.5\% perc): [0.25, 4]
}
\keyword{datasets}
