% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testcase4}
\alias{testcase4}
\title{Simulated Testcase 4.}
\format{
Numeric vector with data points.
}
\usage{
testcase4
}
\description{
This dataset consists of N = 100,000 simulated measurements with 90\\% non-pathological and 10\\% pathological samples.
Ground Truth for reference intervals (2.5\\% perc., 97.5\\% perc): \link{10, 50}
}
\keyword{datasets}
